+++ src/main.js
@@ -1,6 +1,6 @@
 import loginRaizen from './pages/teamRaizen/loginRaizen.js';
 import teamRaizen from './pages/teamRaizen/teamRaizen.js';
-import finalCostumer from './pages/finalCostumer/finalCostumer.js';
+//import finalCostumer from './pages/finalCostumer/finalCostumer.js';
 import initialCostumer from './pages/finalCostumer/initialCostumer.js';
 import searchCompany from './pages/searchCompany/searchCompany.js';
 
@@ -13,12 +13,9 @@ const routes = () => {
       case '#initialCostumer':
         main.appendChild(initialCostumer());
         break;
-      case '#finalCostumer':
-        main.appendChild(finalCostumer());
+      case '#searchCompany':
+        main.appendChild(searchCompany());
         break;
-        case '#searchCompany':
-          main.appendChild(searchCompany());
-          break;
       case '#teamRaizen':
         main.appendChild(teamRaizen());
         break;
