+++ src/pages/finalCostumer/initialCostumer.js
@@ -14,25 +14,22 @@ export default () => {
     <section class="info-lote">
         <input class="input-lote" id="lote-sugar" type="text" placeholder="Digite o número do Lote">
         <button id="btn-lote">PESQUISAR</button>
+        <div id='printLote'></div>
         <p id="error"></p>
-    </section>
     </section>`;
 
     initialPage.innerHTML = templateLote;
 
+    const listLote = initialPage.querySelector('#printLote');
+
     const inputLote = initialPage.querySelector('#lote-sugar');
+    console.log(inputLote);
     const btnSearch = initialPage.querySelector('#btn-lote');
-    //const errorLote = initialPage.querySelector('#error');
-    //btnSearch.addEventListener('click', getLote(products, inputLote.value));
+    console.log(btnSearch);
     btnSearch.addEventListener('click', () => {
-        getLote(products, inputLote.value)
-        window.location.hash = '#finalCostumer';
-        let thisPage = new URL(window.location.href);
-        thisPage.searchParams.append('yourKey', 'someValue');
-        console.log(thisPage);
+        const getLoteArray = getLote(products, inputLote.value)
+        listLote.innerHTML = createCard(getLoteArray);
     });
-
-
 
     return initialPage;
 }
@@ -40,4 +37,23 @@ export default () => {
 function getLote(products, lote) {
     const arrayLote = products.filter((product) => product.lote === lote);
     return arrayLote;
+}
+
+function createCard(products) {
+    const arrayProduct = products.map((product) => {
+        const template = `
+         <section class='sectionCard'>
+            <img class="product-img" src="${product.image}" alt="açúcar">
+            <div class='textCard'><strong>Lote:</strong> ${product.lote}</div>
+            <div class='textCard'><strong>Origem:</strong> ${product.origin}</div>
+            <div class='textCard'><strong>Fazenda:</strong> ${product.farm}</div>
+            <div class='textCard'><strong>Rastreabilidade:</strong> ${product.traceability}</div>
+            <p><strong>Certificações:</strong></p>
+            <div><img class="certification-img" src="${product.certification}" alt="Certificado"></div>
+         </section>
+     `;
+
+        return template;
+    });
+    return arrayProduct.join('');
 }
